Move listwatch package to top-level pkg directory
This moves the `listwatch` package from `security/pkg` to the
top-level `pkg` directory in preparation for using it elsewhere.
